requirements.txt
aiohttp==3.7.4.post0
google-api-python-client==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
pyserial==3.5

# Testing
ipython==7.21.0
hypothesis==6.8.1
pytest==6.2.2

